Authentication System/backend/requirements.txt
flask
flask-cors
flask-jwt-extended
bcrypt
mysql-connector-python
python-dotenv